requirements.txt
attrs==21.4.0
aws-cdk-lib==2.24.0
cattrs==22.1.0
constructs==10.1.41
exceptiongroup==1.0.0rc8
jsii==1.60.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.2.0
